refactored ui entirely
closes #9
