+++ src/ui/CPUUsage/elements.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const Usage = styled.div`
+  position: absolute;
+  bottom: -18px;
+  right: 0px;
+  color: white;
+  font-size: 10px;
+`
