+++ src/ui/LCD/index.js
@@ -1,11 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import PatchSelector from '../PatchSelector'
 import { Wrapper, Container, LCDBorderSide, LCDBorderTop } from './elements'
 
 LCD.propTypes = {
-  patches: PropTypes.arrayOf(PropTypes.object).isRequired,
-  setPatch: PropTypes.func.isRequired,
   children: PropTypes.node.isRequired
 }
 
@@ -14,7 +11,6 @@ export default function LCD({ patches, setPatch, children }) {
     <Wrapper>
       <LCDBorderTop />
       <Container>
-        <PatchSelector patches={patches} setPatch={setPatch} />
         {children}
         <LCDBorderSide />
       </Container>
