+++ src/ui/Piano/index.js
@@ -1,0 +1,24 @@
+import React from 'react'
+import { Piano, KeyboardShortcuts, MidiNumbers } from 'react-piano'
+
+export default React.memo(function PianoWrapper({ noteOn, noteOff, octave }) {
+  const firstNote = MidiNumbers.fromNote('c3')
+  const lastNote = MidiNumbers.fromNote('c5')
+  const keyboardShortcuts = KeyboardShortcuts.create({
+    firstNote: firstNote,
+    lastNote: lastNote,
+    keyboardConfig: [
+      ...KeyboardShortcuts.BOTTOM_ROW,
+      ...KeyboardShortcuts.QWERTY_ROW
+    ]
+  })
+  return (
+    <Piano
+      noteRange={{ first: firstNote, last: lastNote }}
+      playNote={note => noteOn(note + octave)}
+      stopNote={note => noteOff(note + octave)}
+      width={850}
+      keyboardShortcuts={keyboardShortcuts}
+    />
+  )
+})
