+++ src/ui/Record/elements.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+import { FixedSizeButton } from '../Button'
+
+export const RedButton = styled(FixedSizeButton)`
+  color: red;
+`
