+++ src/ui/Splashscreen/elements.js
