+++ src/ui/Splashscreen/index.js
